+++ src/accounts.js
@@ -1,5 +1,5 @@
 export function sortAccounts(accounts) {
-	accounts.value.sort((a, b) => {
+	accounts.sort((a, b) => {
 		const va = `${a.localPart}@${a.domain}`;
 		const vb = `${b.localPart}@${b.domain}`;
 		return va.localeCompare(vb);